+++ sympy/physics/mechanics/tests/test_system_class.py
@@ -4,10 +4,10 @@ from sympy.core.backend import (ImmutableMatrix, symbols, sin, cos, zeros, eye,
                                 _simplify_matrix, sympify)
 from sympy.simplify.simplify import simplify
 from sympy.solvers.solvers import solve
-
+from sympy.physics.mechanics._system import System
 from sympy.physics.mechanics import (
-    System, ReferenceFrame, Point, RigidBody, Particle, dynamicsymbols,
-    PinJoint, PrismaticJoint, Force, Torque, LagrangesMethod, KanesMethod)
+    ReferenceFrame, Point, RigidBody, Particle, dynamicsymbols, PinJoint,
+    PrismaticJoint, Force, Torque, LagrangesMethod, KanesMethod)
 
 t = dynamicsymbols._t  # type: ignore
 q = dynamicsymbols('q:6')  # type: ignore
